requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Async Support
asyncio==3.4.3
asyncpg==0.29.0
aiohttp==3.9.1

# LangChain Ecosystem
langchain==0.1.6
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langgraph==0.0.29
langsmith==0.0.87

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Vector Store
qdrant-client==1.7.0

# Message Queue
google-cloud-pubsub==2.19.0

# Slack Integration
slack-bolt==1.18.1
slack-sdk==3.23.0

# API Framework
fastapi==0.108.0
uvicorn==0.27.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
prometheus-client==0.19.0

# Additional utilities
redis==5.0.1
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0
